RAG_Real_life_use_case_Week3_Day5/playstation-rag-suryansh/requirements.txt
streamlit
faiss-cpu>=1.11.0
sentence-transformers
PyPDF2
groq
numpy
python-dotenv
